prepare dart api

+++ src/main/java/com/example/demo/vo/BalanceSheetDto.java
@@ -8,19 +8,19 @@ public class BalanceSheetDto {
 	private Date reportingYear;		 	//회계연도
 	private String reportCode;			//리포트 코드 (1q, 2q, 3,q 4q)
 	private String fsNm;				//연결재무제표 or 재무제표
-	private int revenue;				//매출
-	private int operatinIncome;			//영업이익
-	private int netIncome;				//당기순이익
-	private int asset;					//자산
-	private int debt;					//부채
-	private int equity;					//자본
-	private int currentAsset;			//단기자산(현금성자산)
-	private int totalNonCurrentAsset;	//장기자산(설비, 부동산 등)
-	private int shortTermDebt;			//단기 부채
-	private int longTermDebt;			//장기 부채
-	private int OCF;					//영업현금흐름
-	private int ICF;					//투자현금흐름
-	private int FCF;					//재무현금흐름
+	private double revenue;				//매출
+	private double operatinIncome;			//영업이익
+	private double netIncome;				//당기순이익
+	private double asset;					//자산
+	private double debt;					//부채
+	private double equity;					//자본
+	private double currentAsset;			//단기자산(현금성자산)
+	private double totalNonCurrentAsset;	//장기자산(설비, 부동산 등)
+	private double shortTermDebt;			//단기 부채
+	private double longTermDebt;			//장기 부채
+	private double OCF;					//영업현금흐름
+	private double ICF;					//투자현금흐름
+	private double FCF;					//재무현금흐름
 	
 	public BalanceSheetDto() {
 		super();
@@ -59,82 +59,82 @@ public class BalanceSheetDto {
 	public void setReportCode(String reportCode) {
 		this.reportCode = reportCode;
 	}
-	public int getRevenue() {
+	public double getRevenue() {
 		return revenue;
 	}
-	public void setRevenue(int revenue) {
+	public void setRevenue(double revenue) {
 		this.revenue = revenue;
 	}
-	public int getOperatinIncome() {
+	public double getOperatinIncome() {
 		return operatinIncome;
 	}
-	public void setOperatinIncome(int operatinIncome) {
+	public void setOperatinIncome(double operatinIncome) {
 		this.operatinIncome = operatinIncome;
 	}
-	public int getNetIncome() {
+	public double getNetIncome() {
 		return netIncome;
 	}
-	public void setNetIncome(int netIncome) {
+	public void setNetIncome(double netIncome) {
 		this.netIncome = netIncome;
 	}
-	public int getAsset() {
+	public double getAsset() {
 		return asset;
 	}
-	public void setAsset(int asset) {
+	public void setAsset(double asset) {
 		this.asset = asset;
 	}
-	public int getDebt() {
+	public double getDebt() {
 		return debt;
 	}
-	public void setDebt(int debt) {
+	public void setDebt(double debt) {
 		this.debt = debt;
 	}
-	public int getEquity() {
+	public double getEquity() {
 		return equity;
 	}
-	public void setEquity(int equity) {
+	public void setEquity(double equity) {
 		this.equity = equity;
 	}
-	public int getCurrentAsset() {
+	public double getCurrentAsset() {
 		return currentAsset;
 	}
-	public void setCurrentAsset(int currentAsset) {
+	public void setCurrentAsset(double currentAsset) {
 		this.currentAsset = currentAsset;
 	}
-	public int getTotalNonCurrentAsset() {
+	public double getTotalNonCurrentAsset() {
 		return totalNonCurrentAsset;
 	}
-	public void setTotalNonCurrentAsset(int totalNonCurrentAsset) {
+	public void setTotalNonCurrentAsset(double totalNonCurrentAsset) {
 		this.totalNonCurrentAsset = totalNonCurrentAsset;
 	}
-	public int getShortTermDebt() {
+	public double getShortTermDebt() {
 		return shortTermDebt;
 	}
-	public void setShortTermDebt(int shortTermDebt) {
+	public void setShortTermDebt(double shortTermDebt) {
 		this.shortTermDebt = shortTermDebt;
 	}
-	public int getLongTermDebt() {
+	public double getLongTermDebt() {
 		return longTermDebt;
 	}
-	public void setLongTermDebt(int longTermDebt) {
+	public void setLongTermDebt(double longTermDebt) {
 		this.longTermDebt = longTermDebt;
 	}
-	public int getOCF() {
+	public double getOCF() {
 		return OCF;
 	}
-	public void setOCF(int oCF) {
+	public void setOCF(double oCF) {
 		OCF = oCF;
 	}
-	public int getICF() {
+	public double getICF() {
 		return ICF;
 	}
-	public void setICF(int iCF) {
+	public void setICF(double iCF) {
 		ICF = iCF;
 	}
-	public int getFCF() {
+	public double getFCF() {
 		return FCF;
 	}
-	public void setFCF(int fCF) {
+	public void setFCF(double fCF) {
 		FCF = fCF;
 	}
